snapshot of refactoring to get two tower candidate generator ready
