+++ src/app/lib/rankers/two_tower.py
@@ -7,64 +7,26 @@ Performs vector-matrix multiplication to get scores for each post, and returns t
 
 import asyncio
 import logging
-import os
 
 from ...models import RankedCandidate, CandidatePost, RankPredictResult
 from .base import Ranker, RankerExecutionError, RankerResult
-from ..elasticsearch import fetch_recent_liked_post_uris, fetch_post_embeddings_and_authors
+from ..elasticsearch import fetch_post_embeddings_and_authors
 from ..embeddings import decode_float32_b64
-from ..feed_debug import current_recorder
 from ..http_client import get_http_client
-from ..request_context import get_request_id
 from ..telemetry import timed
+from ..inference import (
+    get_inference_settings,
+    build_inference_headers,
+    raise_inference_response_error,
+    compute_user_embedding,
+)
 
 
 logger = logging.getLogger(__name__)
 TWO_TOWER_MODEL_NAME = "two_tower"
 
-def get_inference_settings() -> tuple[str, str]:
-    """Load inference configuration only when the two-tower ranker is used."""
-    base_url = os.environ.get("GE_INFERENCE_BASE_URL", "").rstrip("/")
-    if not base_url:
-        raise RankerExecutionError(
-            TWO_TOWER_MODEL_NAME,
-            "GE_INFERENCE_BASE_URL environment variable is required",
-        )
-
-    api_key = os.environ.get("GE_INFERENCE_API_KEY")
-    if not api_key:
-        raise RankerExecutionError(
-            TWO_TOWER_MODEL_NAME,
-            "GE_INFERENCE_API_KEY environment variable is required",
-        )
-
-    return base_url, api_key
-
 
 POST_TOWER_BATCH_SIZE = 32
-
-
-def _build_inference_headers(api_key: str) -> dict[str, str]:
-    """Outbound headers for inference HTTP calls.
-
-    Includes the current request ID (when set) so the inference service
-    can log it alongside our own logs for cross-service correlation.
-    """
-    headers = {"X-API-Key": api_key}
-    rid = get_request_id()
-    if rid is not None:
-        headers["x-request-id"] = rid
-    return headers
-
-
-def _raise_inference_response_error(model_type: str, status_code: int, body: str) -> None:
-    body = body.strip()
-    if len(body) > 2000:
-        body = f"{body[:2000]}..."
-    raise RankerExecutionError(
-        TWO_TOWER_MODEL_NAME,
-        f"{model_type} inference failed status={status_code} body={body}",
-    )
 
 
 async def predict_post_tower_batch(
@@ -75,7 +37,7 @@ async def predict_post_tower_batch(
     api_key: str,
 ) -> list[list[float]]:
     url = f"{base_url}/models/post-tower/predict"
-    headers = _build_inference_headers(api_key)
+    headers = build_inference_headers(api_key)
 
     if len(post_embeddings) != len(author_dids):
         raise RankerExecutionError(
@@ -106,7 +68,7 @@ async def predict_post_tower_batch(
                 resp.status_code,
                 resp.text,
             )
-            _raise_inference_response_error("post-tower", resp.status_code, resp.text)
+            raise_inference_response_error("post-tower", resp.status_code, resp.text)
         return resp.json()["outputs"]
 
     async with timed(
@@ -114,33 +76,6 @@ async def predict_post_tower_batch(
     ):
         chunk_outputs = await asyncio.gather(*(_call_chunk(c) for c in chunks))
     return [item for chunk_out in chunk_outputs for item in chunk_out]
-
-
-async def predict_user_tower_single(
-    history_embeddings: list[list[float]],
-    history_author_dids: list[str],
-    *,
-    base_url: str,
-    api_key: str,
-) -> list[list[float]]:
-    url = f"{base_url}/models/user-tower/predict"
-    headers = _build_inference_headers(api_key)
-    payload = {
-        "history_embeddings": history_embeddings,
-        "history_author_dids": history_author_dids,
-    }
-
-    client = get_http_client()
-    async with timed(logger, "user_tower_http", n_history=len(history_embeddings)):
-        resp = await client.post(url, json=payload, headers=headers)
-    if resp.is_error:
-        logger.error(
-            "user-tower predict failed status=%s body=%s",
-            resp.status_code,
-            resp.text,
-        )
-        _raise_inference_response_error("user-tower", resp.status_code, resp.text)
-    return resp.json()["outputs"]
 
 
 class TwoTowerRanker(Ranker):
@@ -169,49 +104,6 @@ class TwoTowerRanker(Ranker):
 
         valid_candidates = [candidate for candidate in candidates if candidate.at_uri is not None]
         candidates_by_uri = {candidate.at_uri: candidate for candidate in candidates if candidate.at_uri is not None}
-
-        async def _compute_user_embedding() -> list[float]:
-            async with timed(logger, "two_tower_user_side", user_did=user_did):
-                user_history_vectors: list[list[float]] = []
-                history_author_dids: list[str] = []
-                user_history_liked_uris = await fetch_recent_liked_post_uris(es, user_did)
-
-                rec = current_recorder()
-
-                if not user_history_liked_uris:
-                    logger.info("No likes found for user %s", user_did)
-                    if rec is not None:
-                        rec.record_user_features(self.name, [], 0)
-                else:
-                    user_history_embedding_pairs: list[tuple[str, list[float], str]] = await fetch_post_embeddings_and_authors(
-                        es, user_history_liked_uris,
-                    )
-                    if rec is not None:
-                        rec.record_user_features(
-                            self.name, user_history_liked_uris, len(user_history_embedding_pairs)
-                        )
-                    if not user_history_embedding_pairs:
-                        logger.info(
-                            "No embeddings found for %d liked posts of user %s",
-                            len(user_history_liked_uris),
-                            user_did,
-                        )
-                    else:
-                        user_history_vectors = [embedding for _, embedding, _ in user_history_embedding_pairs]
-                        history_author_dids = [author_did for _, _, author_did in user_history_embedding_pairs]
-
-                output_user_embedding_list = await predict_user_tower_single(
-                    user_history_vectors,
-                    history_author_dids,
-                    base_url=inference_base_url,
-                    api_key=inference_api_key,
-                )
-                if len(output_user_embedding_list) != 1:
-                    raise RankerExecutionError(
-                        self.name,
-                        f"user inference returned {len(output_user_embedding_list)} embeddings; expected 1",
-                    )
-                return output_user_embedding_list[0]
 
         async def _compute_candidate_post_embeddings() -> (
             tuple[list[CandidatePost], list[list[float]]] | None
@@ -265,7 +157,13 @@ class TwoTowerRanker(Ranker):
                 return ranked_candidates_input, output_post_embeddings
 
         output_user_embedding, candidate_result = await asyncio.gather(
-            _compute_user_embedding(),
+            compute_user_embedding(
+                user_did,
+                es,
+                inference_base_url,
+                inference_api_key,
+                TWO_TOWER_MODEL_NAME
+            ),
             _compute_candidate_post_embeddings(),
         )
